backend/requirements.txt
# Merged from backend/requirements.txt and backend/case/requirements.txt
# Conflict resolution: prefer newer/more specific pinned versions when present.
# uvicorn kept with [standard] extras and upgraded to 0.38.0 to match other file.

annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
easyocr>=1.7.1
fastapi==0.120.3
greenlet==3.2.4
h11==0.16.0
httpx==0.27.2
idna==3.11
openai==1.51.0
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.2.1
python-multipart>=0.0.9
pymupdf>=1.24.9
requests>=2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn[standard]==0.38.0
bcrypt==4.2.0

